pyproject.toml
[project]
name = "compress-and-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "datasets==2.16.1",
    "hydra-core>=1.3.2",
    "mlflow>=3.1.0",
    "numpy==1.21.4",
    "onnx>=1.17.0",
    "opt-einsum==3.3.0",
    "pytorch-lightning>=2.3.3",
    "scipy==1.8.0",
    "torch==2.0.0",
    "transformers==4.37.1",
]

[dependency-groups]
dev = [
    "fire>=0.7.0",
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "rich>=14.0.0",
    "wandb>=0.20.1",
]

[tool.black]
    line-length = 88
    target-version = ["py310"]

[tool.isort]
    profile = "black"
    line_length = 88
    multi_line_output = 3
    include_trailing_comma = true
    force_grid_wrap = 0
    use_parentheses = true

[tool.flake8]
    max-line-length = 88
    extend-ignore = ["E203", "W503"]  # black несовместим с этими
